Add files via upload
The car assembly line is a classic example of a multi-station (multi-stage) pipeline processing in
concurrency that employs a producer-consumer synchronization pattern between consecutive
stations along the pipeline. The current station assembles the work in progress (WIP) vehicle
conveyed from the previous station (if there is any) and transfers it to the next station for
further assembling, the process stops at the end of line (EOL). The current station is the
consumer of the WIP vehicle that passed from the previous station, while it is also the producer
of the WIP vehicle to the next station.